requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
billiard==3.6.4.0
    # via celery
black==21.9b0
    # via -r requirements.in
boto3==1.18.58
    # via django-s3-folder-storage
botocore==1.21.58
    # via
    #   boto3
    #   s3transfer
celery==5.1.2
    # via
    #   -r requirements.in
    #   django-celery-beat
click==7.1.2
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
collectfast==2.2.0
    # via -r requirements.in
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.8
    # via
    #   -r requirements.in
    #   collectfast
    #   django-celery-beat
    #   django-dbbackup
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-logentry-admin
    #   django-role-permissions
    #   django-storages
    #   django-timezone-field
django-admin-rangefilter==0.8.1
    # via -r requirements.in
django-celery-beat==2.2.1
    # via -r requirements.in
django-dbbackup==3.3.0
    # via -r requirements.in
django-debug-toolbar==3.2.2
    # via -r requirements.in
django-discover-runner==1.0
    # via django-role-permissions
django-extensions==3.1.3
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-logentry-admin==1.1.0
    # via -r requirements.in
django-role-permissions==3.1.1
    # via -r requirements.in
django-s3-folder-storage==0.5
    # via -r requirements.in
django-storages==1.12
    # via
    #   collectfast
    #   django-s3-folder-storage
django-timezone-field==4.2.1
    # via django-celery-beat
gunicorn==20.1.0
    # via -r requirements.in
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsondate==0.1.2
    # via -r requirements.in
kombu==5.1.0
    # via celery
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via black
prompt-toolkit==3.0.20
    # via click-repl
psycopg2-binary==2.9.1
    # via -r requirements.in
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   python-crontab
python-decouple==3.5
    # via -r requirements.in
pytz==2021.3
    # via
    #   -r requirements.in
    #   celery
    #   django
    #   django-dbbackup
    #   django-timezone-field
raven==6.10.0
    # via -r requirements.in
redis==3.5.3
    # via -r requirements.in
regex==2021.10.8
    # via black
s3transfer==0.5.0
    # via boto3
six==1.16.0
    # via
    #   click-repl
    #   django-dbbackup
    #   django-role-permissions
    #   python-dateutil
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
tomli==1.2.1
    # via black
typing-extensions==3.10.0.2
    # via
    #   black
    #   collectfast
urllib3==1.26.7
    # via botocore
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
